bugfix: Fixed joining StringLists when last node has child
When calling pop_node() on a StringList, we create a new list that
uses the same StringNodes as a previous list but the pointer to the
last node points to the second to last node in the original list.
As a result, when joining those types of StringLists we didn't stop
at the last node for the list but we kept going until there were no
more nodes left from the original list.